LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/clipped-scrollers-skipped-expected.txt


FAIL Anchor selection with nested scrollers. assert_equals: expected 300 but got 100
